[refactor](catalog) move some package related to catalog
