+++ fe/fe-core/src/main/java/org/apache/doris/catalog/EsResource.java
@@ -19,7 +19,7 @@ package org.apache.doris.catalog;
 
 import org.apache.doris.common.DdlException;
 import org.apache.doris.common.proc.BaseProcResult;
-import org.apache.doris.external.elasticsearch.EsUtil;
+import org.apache.doris.datasource.es.EsUtil;
 
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
