+++ fe/fe-core/src/main/java/org/apache/doris/common/proc/EsShardProcDir.java
@@ -21,8 +21,8 @@ import org.apache.doris.catalog.Database;
 import org.apache.doris.catalog.EsTable;
 import org.apache.doris.common.AnalysisException;
 import org.apache.doris.common.util.ListComparator;
-import org.apache.doris.external.elasticsearch.EsShardPartitions;
-import org.apache.doris.external.elasticsearch.EsShardRouting;
+import org.apache.doris.datasource.es.EsShardPartitions;
+import org.apache.doris.datasource.es.EsShardRouting;
 
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
