+++ fe/fe-core/src/main/java/org/apache/doris/datasource/es/MappingPhase.java
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package org.apache.doris.external.elasticsearch;
+package org.apache.doris.datasource.es;
 
 import org.apache.doris.catalog.Column;
 import org.apache.doris.catalog.EsTable;
