+++ fe/fe-core/src/main/java/org/apache/doris/datasource/iceberg/IcebergUtils.java
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package org.apache.doris.external.iceberg.util;
+package org.apache.doris.datasource.iceberg;
 
 
 import org.apache.doris.analysis.BinaryPredicate;
