+++ fe/fe-core/src/test/java/org/apache/doris/external/elasticsearch/PartitionPhaseTest.java
@@ -21,6 +21,11 @@ import org.apache.doris.catalog.Column;
 import org.apache.doris.catalog.EsTable;
 import org.apache.doris.catalog.PrimitiveType;
 import org.apache.doris.common.ExceptionChecker;
+import org.apache.doris.datasource.es.EsNodeInfo;
+import org.apache.doris.datasource.es.EsRestClient;
+import org.apache.doris.datasource.es.EsShardPartitions;
+import org.apache.doris.datasource.es.PartitionPhase;
+import org.apache.doris.datasource.es.SearchContext;
 
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.ObjectMapper;
